// src/store/action-types/title-action-types.ts
export enum TitleActionTypes{
    SET_TITLES = "SET_TITLES",
    SET_TITLE_ERROR = "SET_TITLE_ERROR",
    SET_RANDOM_TITLES = "SET_RANDOM_TITLE",
    SET_PAGINATED_TITLES = "SET_PAGINATED_TITLES",
    SET_LAST_TITLE_PAGE = "SET_LAST_TITLE_PAGE",
    SET_IS_TITLE = "SET_IS_TITLE",
    SET_TITLE_SEARCH_RESULT = "SET_TITLE_SEARCH_RESULT",
    SET_TITLE_SEARCH_VALUE = "SET_TITLE_SEARCH_VALUE",
    SET_TITLE_GENRES = "SET_TITLE_GENRES",
    SET_TITLE_BY_GENRE = "SET_TITLE_BY_GENRE",
    SET_CURRENT_TITLE_TITLE = "SET_CURRENT_TITLE_TITLE",
    SET_CURRENT_TITLE_PAGE = "SET_CURRENT_TITLE_PAGE",
    SET_POPULAR_TITLE = "SET_POPULAR_TITLE",
    SET_TITLE_REVIEW = "SET_TITLE_REVIEW",
    SET_TITLE_CLEANUP_RANDOM_TITLES = "SET_TITLE_CLEANUP_RANDOM_TITLES",
    SET_TITLE_CLEANUP_PAGINATED_TITLES = "SET_TITLE_CLEANUP_PAGINATED_TITLES",
    SET_TITLE_CLEANUP_SEARCH_RESULT = "SET_TITLE_CLEANUP_SEARCH_RESULT",
    SET_TITLE_CLEANUP_TITLE_GENRES = "SET_TITLE_CLEANUP_TITLE_GENRES",
    SET_TITLE_CLEANUP_TITLE_BY_GENRES = "SET_TITLE_CLEANUP_TITLE_BY_GENRES",
    SET_TITLE_CLEANUP_CURRENT_TITLE = "SET_TITLE_CLEANUP_CURRENT_TITLE",
    SET_TITLE_CLEANUP_POPULAR_TITLE = "SET_TITLE_CLEANUP_POPULAR_TITLE",
    SET_TITLE_CLEANUP_TITLE_REVIEWS = "SET_TITLE_CLEANUP_TITLE_REVIEWS",
}